init locals fpga_pci_mbox_scan & fpga_pci_get_dma_device_num
in fpga_pci_mbox_scan there is a path where it ret can get to
fail_with_code() and not be initialized (past 'No more directories'
comment).  Initialize it to an error condition so that uninitialized
value can not be returned.

in fpga_pci_get_dma_device_num, explicitly initialize dirp to NULL
so that the early error case can not cause the error cleanup to
call closedir() on an uninitialized value.